requirements.txt
youtube-transcript-api
langchain
langchain-openai
langchain-community
langchain-google-genai
faiss-cpu
tiktoken
python-dotenv
sentence-transformers
youtube-transcript-api
